% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_palette.R
\docType{data}
\name{esc_deepgreen}
\alias{esc_deepgreen}
\title{Palette generator for esc_deepgreen}
\format{
An object of class \code{character} of length 9.
}
\usage{
esc_deepgreen
}
\description{
Palette generator for esc_deepgreen
}
\keyword{datasets}
