% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_palette.R
\name{esc_palette}
\alias{esc_palette}
\title{Predefined ESC colours combined into palettes}
\usage{
esc_palette(n = 6, reverse = FALSE)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale, defaults to 6. Minimum value is
1, maximum is 12. Using more than 6 colours is not recommended as it may
make it difficult to distinguish between colours.}

\item{reverse}{Logical. Reverse the standard colour order, defaults to FALSE}
}
\description{
Predefined ESC colours combined into palettes
}
