% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_theme.R
\name{esc_theme}
\alias{esc_theme}
\title{Create a ggplot2 theme consistent with the ESC style guide.}
\usage{
esc_theme(scale = 0.6)
}
\arguments{
\item{scale}{Adjust the scale of the chart. Default is 0.6}
}
\description{
Create a ggplot2 theme consistent with the ESC style guide.
}
