% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_theme_sf.R
\name{esc_theme_sf}
\alias{esc_theme_sf}
\title{Create a ggplot2 spatial theme consistent with the ESC style guide.}
\usage{
esc_theme_sf(scale = 0.6)
}
\arguments{
\item{scale}{Adjust the scale of the chart. Default is 0.6}
}
\description{
Create a ggplot2 spatial theme consistent with the ESC style guide.
}
