% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esctheme-package.R
\docType{package}
\name{esctheme-package}
\alias{esctheme}
\alias{esctheme-package}
\title{esctheme: Create ggplot2 charts in the Essential Services Commission of Victoria style}
\description{
Enables the user to create ggplot2 charts that conform to the Essential Services Commission of Victoria style guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://esc-economics.github.io/esctheme/}
}

}
\author{
\strong{Maintainer}: Economics and Insights ESC \email{economics@esc.vic.gov.au}

Authors:
\itemize{
  \item Miguel Ayala \email{miguel.ayala@esc.vic.gov.au}
}

}
\keyword{internal}
