% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{r_setup}
\alias{r_setup}
\title{A general setup function that clears the console and workspace and loads in the packages you need to create ESC style charts}
\usage{
r_setup()
}
\description{
A general setup function that clears the console and workspace and loads in the packages you need to create ESC style charts
}
