% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_functions.R
\name{summarise_by}
\alias{summarise_by}
\title{Calculate distribution statistics
function to calculate median, min, max and lower and upper values}
\usage{
summarise_by(df, var, lower = 0.1, upper = 0.9, by_vars)
}
\arguments{
\item{df}{A data frame}

\item{var}{Numeric column to compute distribution.}

\item{lower}{Numeric value for the bottom percentile.}

\item{upper}{Numeric value for the top percentile.}

\item{by_vars}{A selection of columns to group by for just this operation}
}
\description{
Calculate distribution statistics
function to calculate median, min, max and lower and upper values
}
