% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_functions.R
\name{trail_avg}
\alias{trail_avg}
\title{Moving average for long data}
\usage{
trail_avg(df, p = 4, ivars, by_vars)
}
\arguments{
\item{df}{A data frame}

\item{p}{Numeric value for window period. Default to 4 for quarter.
Other options are weekly 7 and monthly 12}

\item{ivars}{Coumns to calculate moving average. It keeps the column names}

\item{by_vars}{A selection of columns to group by for just this operation}
}
\description{
A function to calculate moving average to smooth time series
}
