% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_functions.R
\name{y_avg_growth}
\alias{y_avg_growth}
\title{Year average growth}
\usage{
y_avg_growth(x, p = 4, ivars = unique(x$variable))
}
\arguments{
\item{x}{data in long format. Data in 3 columns (Dates, variable and value)}

\item{p}{Number of periods to calculate year average growth (defaults to 4)}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}
}
\value{
Transformed data, still in long format
}
\description{
Calculates year average growth rates
}
