% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_functions.R
\name{esc_save}
\alias{esc_save}
\title{generic functions used in the package  ####}
\usage{
esc_save(n, path = NULL, device = "png")
}
\arguments{
\item{n}{Name given to chart}
}
\value{
Saved png file
}
\description{
generic functions used in the package  ####
}
